+++ CHANGELOG.md
@@ -75,6 +75,9 @@ TODO
 - Support for multiple control servers per AS (fallback), compartmentalize GRPC usage and
   fully switch to OkHttp.
   [#179](https://github.com/scionproto-contrib/jpan/pull/179)
+TODO
+- close() IOException?
+- Remove GRPC usage from Segments
 
 ### Fixed
 - Post 0.5.0 cleanup